Squashed commit of the following:
commit b08bcc2ffc04597d06d330fc613a975d82b9dab8
Merge: 0d2a720 f27c1dc
Author: Narayan soni <narayansoni854@gmail.com>
Date:   Thu Mar 2 15:17:16 2023 +0530

    Merge pull request #224 from narayan954/deepsource-transform-b2036ef1

commit f27c1dce39cfac01831a82404bb3cbb248520dc9
Author: deepsource-autofix[bot] <62050782+deepsource-autofix[bot]@users.noreply.github.com>
Date:   Thu Mar 2 09:30:02 2023 +0000

    Format code with prettier

    This commit fixes the style issues introduced in 0d2a720 according to the output
    from prettier.

    Details: https://deepsource.io/gh/narayan954/dummygram/transform/293f1d52-4755-4ea8-be7b-9949010c717b/

commit 0d2a720f4d1fe3592b8179560d9d575b776eecbe
Merge: b9375ae 75c8271
Author: Narayan soni <narayansoni854@gmail.com>
Date:   Thu Mar 2 14:59:39 2023 +0530

    Merge pull request #222 from sarkartanmay393/feature/LoginPages

commit 75c82719f88834a9df7d94dbeb226d6904a05921
Author: Narayan soni <narayansoni854@gmail.com>
Date:   Thu Mar 2 14:40:31 2023 +0530

    fixed signup redirect

commit 0a65ccb36a7244023df829f000f01cc63b6315dd
Author: Tanmay Sarkar <sarkartanmay393@gmail.com>
Date:   Thu Mar 2 12:32:27 2023 +0530

    Fix: minor

commit 57dd0726df1674952dc1739442b19a28764b66eb
Author: Tanmay Sarkar <sarkartanmay393@gmail.com>
Date:   Thu Mar 2 12:08:32 2023 +0530

    Fix: noauth user to LoginScreen

commit 7e837a605a5a828c9607ac6408986c81ceaa9733
Author: Tanmay Sarkar <sarkartanmay393@gmail.com>
Date:   Wed Mar 1 21:47:54 2023 +0530

    Fix: AuthScreen functionality

commit f8e7c8d4ce8bb2ae27d825dde1600076b90a4701
Author: Tanmay Sarkar <sarkartanmay393@gmail.com>
Date:   Wed Mar 1 21:22:57 2023 +0530

    Add: login signup initials

commit b9375ae60330f65c6f6d0fa1d87eeee5d217a744
Merge: 5af2310 18e40cb
Author: Narayan soni <narayansoni854@gmail.com>
Date:   Wed Mar 1 20:26:25 2023 +0530

    Merge pull request #221 from narayan954/deepsource-transform-21500910

commit 18e40cb82afe38f8b251121aa28334319806491b
Author: deepsource-autofix[bot] <62050782+deepsource-autofix[bot]@users.noreply.github.com>
Date:   Wed Mar 1 14:54:34 2023 +0000

    Format code with prettier

    This commit fixes the style issues introduced in 5af2310 according to the output
    from prettier.

    Details: https://deepsource.io/gh/narayan954/dummygram/transform/e84265f7-5492-47b1-b900-461d746f5133/

commit 5af2310a71bf87b42839e3ac35e122b730c5f440
Merge: 637610c 08261da
Author: Narayan soni <narayansoni854@gmail.com>
Date:   Wed Mar 1 20:24:18 2023 +0530

    Merge pull request #218 from sarkartanmay393/feature/DarkMode

commit 08261dae60d38453653fe0c1d4155f99d98b9c41
Author: Tanmay Sarkar <sarkartanmay393@gmail.com>
Date:   Wed Mar 1 14:34:35 2023 +0530

    Fix: theme

commit 430b06d553873dbb617b65e3f99e2e3ba5e7d561
Author: Tanmay Sarkar <sarkartanmay393@gmail.com>
Date:   Wed Mar 1 14:27:46 2023 +0530

    Fix: minors

commit 2369ed781a559114f09d82324513e42d312935d5
Merge: 10aa358 c8fc312
Author: Tanmay Sarkar <sarkartanmay393@gmail.com>
Date:   Wed Mar 1 14:27:06 2023 +0530

    Merge remote-tracking branch 'origin/feature/DarkMode' into feature/DarkMode

commit 10aa35869d16196e6fb588cfbb86210714f8edfd
Author: Tanmay Sarkar <sarkartanmay393@gmail.com>
Date:   Wed Mar 1 14:26:19 2023 +0530

    Fix: Theme

commit c8fc312b338a58480ff7b535629c4a53ec2d330f
Merge: 55e6792 637610c
Author: Narayan soni <narayansoni854@gmail.com>
Date:   Tue Feb 28 23:07:00 2023 +0530

    Merge branch 'master' into pr/218

commit 55e6792af85c6d0099f0f3b58489058bbf2699cd
Author: Tanmay Sarkar <sarkartanmay393@gmail.com>
Date:   Tue Feb 28 13:00:24 2023 +0530

    Add: Simple theme mode
