+++ src/App.jsx
@@ -1,29 +1,19 @@
-import React, { useState, useEffect, useMemo } from "react";
+import React, { useState, useEffect } from "react";
 import Post from "./components/Post";
-import {
-  db,
-  auth,
-  storage,
-  googleProvider,
-  facebookProvider,
-} from "./lib/firebase";
-import {
-  Modal,
-  Button,
-  Input,
-  Dialog,
-  DialogTitle,
-  DialogContent,
-} from "@mui/material";
+import { db, auth } from "./lib/firebase";
+import { Button, Dialog, Modal, DialogContent } from "@mui/material";
 import { makeStyles } from "@mui/styles";
 import ImgUpload from "./components/ImgUpload";
 import Loader from "./components/Loader";
-import AnimatedButton from "./components/AnimatedButton";
 import { FaArrowCircleUp } from "react-icons/fa";
 import { useSnackbar } from "notistack";
 import logo from "./assets/logo.png";
+import { Switch, Route, useHistory } from "react-router-dom";
+import LoginScreen from "./pages/Login";
+import SignupScreen from "./pages/Signup";
+import AnimatedButton from "./components/AnimatedButton";
 
-function getModalStyle() {
+export function getModalStyle() {
   const top = 50;
   const left = 50;
   const padding = 5;
@@ -36,43 +26,35 @@ function getModalStyle() {
     padding: `${padding}%`,
     borderRadius: `${radius}%`,
     textAlign: "center",
+    backgroundColor: "var(--bg-color)",
   };
 }
 
-const useStyles = makeStyles((theme) => ({
+export const useStyles = makeStyles((theme) => ({
   paper: {
     position: "absolute",
     width: 200,
-    backgroundColor: theme.palette.background.paper,
-    border: "2px solid #000",
+    border: "1px solid var(--color)",
     borderRadius: theme.shape.borderRadius,
-    boxShadow: theme.shadows[5],
+    boxShadow: theme.shadows[10],
     padding: theme.spacing(2, 4, 3),
+    color: "var(--color)",
   },
 }));
 
 function App() {
   const classes = useStyles();
 
-  const [modalStyle] = useState(getModalStyle);
+  const history = useHistory();
+
   const [posts, setPosts] = useState([]);
-  const [openSignUp, setOpenSignUp] = useState(false);
-  const [openSignIn, setOpenSignIn] = useState(false);
-  const [username, setUsername] = useState("");
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
   const [user, setUser] = useState(null);
-  const [signingUp, setSigningUp] = useState(false);
-  const [loggingIn, setLoggingIn] = useState(false);
   const [loadingPosts, setLoadingPosts] = useState(true);
   const [pageSize, setPageSize] = useState(10);
   const [loadMorePosts, setLoadMorePosts] = useState(false);
   const [openNewUpload, setOpenNewUpload] = useState(false);
   const [logout, setLogout] = useState(false);
-  const processingAuth = useMemo(
-    () => loggingIn || signingUp || loadingPosts,
-    [loggingIn, signingUp, loadingPosts]
-  );
+
   const buttonStyle = {
     background: "linear-gradient(40deg, #e107c1, #59afc7)",
     borderRadius: "20px",
@@ -81,18 +63,8 @@ function App() {
     },
   };
 
-  const [image, setImage] = useState(null);
-  const [address, setAddress] = useState(null);
-
   const { enqueueSnackbar } = useSnackbar();
   const [showScroll, setShowScroll] = useState(false);
-
-  const handleChange = (e) => {
-    if (e.target.files[0]) {
-      setImage(e.target.files[0]);
-      setAddress(e.target.value);
-    }
-  };
 
   const checkScrollTop = () => {
     if (!showScroll && window.pageYOffset > 400) {
@@ -111,20 +83,34 @@ function App() {
   useEffect(() => {
     const unsubscribe = auth.onAuthStateChanged((authUser) => {
       if (authUser) {
-        // user has logged in
         setUser(authUser);
+        history.push("/dummygram/");
       } else {
-        // user has logged out
         setUser(null);
+        history.push("/dummygram/login");
       }
     });
+
     return () => {
-      // perform some cleanup actions
       unsubscribe();
     };
-  }, [user, username]);
+  }, [user]);
 
   useEffect(() => {
+    if (document.body.classList.contains("darkmode--activated")) {
+      window.document.body.style.setProperty("--bg-color", "black");
+      window.document.body.style.setProperty("--color", "white");
+      window.document.body.style.setProperty("--val", 1);
+      document.getElementsByClassName("app__header__img").item(0).style.filter =
+        "invert(100%)";
+    } else {
+      window.document.body.style.setProperty("--bg-color", "white");
+      window.document.body.style.setProperty("--color", "#2B1B17");
+      window.document.body.style.setProperty("--val", 0);
+      document.getElementsByClassName("app__header__img").item(0).style.filter =
+        "invert(0%)";
+    }
+
     window.addEventListener("scroll", handleMouseScroll);
     db.collection("posts")
       .orderBy("timestamp", "desc")
@@ -170,126 +156,11 @@ function App() {
     setLoadMorePosts(false);
   }, [loadMorePosts]);
 
-  const signUp = (e) => {
-    e.preventDefault();
-    setSigningUp(true);
-    auth
-      .createUserWithEmailAndPassword(email, password)
-      .then((authUser) => {
-        const uploadTask = storage.ref(`images/${image.name}`).put(image);
-        uploadTask.on(
-          "state_changed",
-          (snapshot) => {
-            // // progress function ...
-            // setProgress(Math.round((snapshot.bytesTransferred / snapshot.totalBytes) * 100));
-          },
-          (error) => {
-            // error function ...
-
-            enqueueSnackbar(error.message, {
-              variant: "error",
-            });
-          },
-          () => {
-            // complete function ...
-            storage
-              .ref("images")
-              .child(image.name)
-              .getDownloadURL()
-              .then((url) => {
-                authUser.user.updateProfile({
-                  displayName: username,
-                  photoURL: url,
-                });
-                enqueueSnackbar("Signup Successful!", {
-                  variant: "success",
-                });
-                setOpenSignUp(false);
-              });
-          }
-        );
-      })
-      // .then(() => {
-
-      // })
-      .catch((error) =>
-        enqueueSnackbar(error.message, {
-          variant: "error",
-        })
-      )
-      .finally(() => {
-        setSigningUp(false);
-      });
-  };
-
-  const signIn = (e) => {
-    e.preventDefault();
-    setLoggingIn(true);
-    auth
-      .signInWithEmailAndPassword(email, password)
-      .then(() => {
-        enqueueSnackbar("Login successful!", {
-          variant: "success",
-        });
-        setOpenSignIn(false);
-      })
-      .catch((error) =>
-        enqueueSnackbar(error.message, {
-          variant: "error",
-        })
-      )
-      .finally(() => {
-        setLoggingIn(false);
-      });
-  };
-
   const signOut = () => {
     auth.signOut().finally();
     enqueueSnackbar("Logged out Successfully !", {
       variant: "info",
     });
-  };
-
-  const signInWithGoogle = (e) => {
-    e.preventDefault();
-    setLoggingIn(true);
-    auth
-      .signInWithPopup(googleProvider)
-      .then(() => {
-        enqueueSnackbar("Login successful!", {
-          variant: "success",
-        });
-        setOpenSignIn(false);
-      })
-      .catch((error) =>
-        enqueueSnackbar(error.message, {
-          variant: "error",
-        })
-      )
-      .finally(() => {
-        setLoggingIn(false);
-      });
-  };
-
-  const signInWithFacebook = (e) => {
-    e.preventDefault();
-    setLoggingIn(true);
-    auth
-      .signInWithPopup(facebookProvider)
-      .then(() => {
-        enqueueSnackbar("Login successful!", {
-          variant: "success",
-        });
-        setOpenSignIn(false);
-      })
-      .catch((error) =>
-        enqueueSnackbar(error.message, {
-          variant: "error",
-        })
-      )
-      .finally(() => {
-        setLoggingIn(false);
-      });
   };
 
   return (
@@ -301,6 +172,10 @@ function App() {
           className="app__header__img w-100"
           onClick={() => {
             window.scrollTo({ top: 0, left: 0, behavior: "smooth" });
+            window.location.href = "/";
+          }}
+          style={{
+            cursor: "pointer",
           }}
         />
         {user ? (
@@ -327,7 +202,9 @@ function App() {
         ) : (
           <div className="login__container">
             <Button
-              onClick={() => setOpenSignIn(true)}
+              onClick={() => {
+                history.push("/dummygram/login");
+              }}
               color="primary"
               variant="contained"
               style={{ margin: 5 }}
@@ -337,7 +214,9 @@ function App() {
             </Button>
 
             <Button
-              onClick={() => setOpenSignUp(true)}
+              onClick={() => {
+                history.push("/dummygram/signup");
+              }}
               color="primary"
               variant="contained"
               style={{ margin: 5 }}
@@ -348,22 +227,44 @@ function App() {
           </div>
         )}
       </div>
+
       <Dialog
         sx={{ borderRadius: "100px" }}
         open={openNewUpload}
         onClose={() => setOpenNewUpload(false)}
       >
         <div
-          style={{ padding: "20px", borderRadius: "10%", textAlign: "center" }}
+          style={{
+            backgroundColor: "var(--bg-color)",
+            padding: "20px",
+            textAlign: "center",
+            color: "var(--color)",
+            border: "2px solid var(--color)",
+          }}
         >
           <img
             src="https://user-images.githubusercontent.com/27727921/185767526-a002a17d-c12e-4a6a-82a4-dd1a13a5ecda.png"
             alt="instagram"
             className="modal__signup__img"
-            style={{ width: "50%" }}
+            style={{ width: "50%", filter: "invert(var(--val))" }}
           />
-          <p style={{ fontSize: "25px", fontFamily: "monospace" }}>New Post</p>
-          <DialogContent>
+          <p
+            style={{
+              fontSize: "25px",
+              fontFamily: "monospace",
+              color: "var(--color)",
+            }}
+          >
+            New Post
+          </p>
+
+          <DialogContent
+            sx={
+              {
+                // backgroundColor: "var(--bg-color)",
+              }
+            }
+          >
             {!loadingPosts &&
               (user ? (
                 <ImgUpload
@@ -376,139 +277,6 @@ function App() {
           </DialogContent>
         </div>
       </Dialog>
-
-      <Modal open={openSignUp} onClose={() => setOpenSignUp(false)}>
-        <div style={modalStyle} className={classes.paper}>
-          <form className="modal__signup" onSubmit={signUp}>
-            <img
-              src="https://user-images.githubusercontent.com/27727921/185767526-a002a17d-c12e-4a6a-82a4-dd1a13a5ecda.png"
-              alt="instagram"
-              className="modal__signup__img"
-              style={{ width: "80%", marginLeft: "10%" }}
-            />
-            <div
-              style={{
-                height: "100px",
-                width: "100px",
-                borderRadius: "100%",
-                border: "2px",
-                borderColor: "black",
-                borderStyle: "solid",
-                marginLeft: "22%",
-              }}
-            >
-              {address ? (
-                <img
-                  src={URL.createObjectURL(image)}
-                  alt="profile pic"
-                  style={{
-                    width: "100px",
-                    height: "100px",
-                    borderRadius: "100%",
-                  }}
-                />
-              ) : (
-                <div style={{ marginTop: "30px" }}>PROFILE PICTURE</div>
-              )}
-            </div>
-            <Input
-              type="text"
-              placeholder="USERNAME"
-              required
-              value={username}
-              style={{ margin: "5%" }}
-              onChange={(e) => setUsername(e.target.value)}
-            />
-            <Input
-              type="text"
-              placeholder="EMAIL"
-              value={email}
-              style={{ margin: "5%" }}
-              onChange={(e) => setEmail(e.target.value)}
-            />
-            <Input
-              type="password"
-              placeholder="PASSWORD"
-              value={password}
-              style={{ margin: "5%" }}
-              onChange={(e) => setPassword(e.target.value)}
-            />
-            <div className="file-input">
-              <input
-                type="file"
-                id="file"
-                className="file"
-                onChange={handleChange}
-                accept="image/*"
-              />
-              <label htmlFor="file">Select Profile Picture</label>
-            </div>
-            <AnimatedButton
-              type="submit"
-              variant="contained"
-              color="primary"
-              sx={buttonStyle}
-            >
-              Sign Up
-            </AnimatedButton>
-          </form>
-        </div>
-      </Modal>
-
-      <Modal open={openSignIn} onClose={() => setOpenSignIn(false)}>
-        <div style={getModalStyle()} className={classes.paper}>
-          <form className="modal__signup">
-            <img
-              src="https://user-images.githubusercontent.com/27727921/185767526-a002a17d-c12e-4a6a-82a4-dd1a13a5ecda.png"
-              alt="dummygram"
-              className="modal__signup__img"
-              style={{ width: "80%", marginLeft: "10%" }}
-            />
-            <Input
-              type="text"
-              placeholder="EMAIL"
-              value={email}
-              style={{ margin: "5%" }}
-              onChange={(e) => setEmail(e.target.value)}
-            />
-            <Input
-              type="password"
-              placeholder="PASSWORD"
-              value={password}
-              style={{ margin: "5%" }}
-              onChange={(e) => setPassword(e.target.value)}
-            />
-            <AnimatedButton
-              type="submit"
-              onClick={signIn}
-              variant="contained"
-              color="primary"
-              sx={buttonStyle}
-            >
-              Sign In
-            </AnimatedButton>
-            <AnimatedButton
-              type="submit"
-              onClick={signInWithGoogle}
-              variant="contained"
-              color="primary"
-              sx={buttonStyle}
-            >
-              Sign In With Google
-            </AnimatedButton>
-            <AnimatedButton
-              type="submit"
-              onClick={signInWithFacebook}
-              variant="contained"
-              color="primary"
-              sx={buttonStyle}
-            >
-              Sign In With Facebook
-            </AnimatedButton>
-          </form>
-        </div>
-      </Modal>
-
       <Modal open={logout} onClose={() => setLogout(false)}>
         <div style={getModalStyle()} className={classes.paper}>
           <form className="modal__signup">
@@ -516,7 +284,11 @@ function App() {
               src="https://user-images.githubusercontent.com/27727921/185767526-a002a17d-c12e-4a6a-82a4-dd1a13a5ecda.png"
               alt="dummygram"
               className="modal__signup__img"
-              style={{ width: "80%", marginLeft: "10%" }}
+              style={{
+                width: "80%",
+                marginLeft: "10%",
+                filter: "invert(var(--val))",
+              }}
             />
 
             <p
@@ -524,6 +296,7 @@ function App() {
                 fontSize: "15px",
                 fontFamily: "monospace",
                 padding: "10%",
+                color: "var(--color)",
               }}
             >
               Are you sure you want to Logout?
@@ -542,37 +315,60 @@ function App() {
         </div>
       </Modal>
 
-      <div
-        style={{
-          display: "flex",
-          alignContent: "center",
-          justifyContent: "center",
-        }}
-      >
-        <div
-          style={
-            !loadingPosts
-              ? {}
-              : {
-                  width: "100%",
-                  minHeight: "100vh",
-                  display: "flex",
-                  justifyContent: "center",
-                  alignItems: "center",
+      <Switch>
+        <Route exact path="/dummygram/">
+          {user ? (
+            <div
+              style={{
+                display: "flex",
+                alignContent: "center",
+                justifyContent: "center",
+              }}
+            >
+              <div
+                style={
+                  !loadingPosts
+                    ? {}
+                    : {
+                        width: "100%",
+                        minHeight: "100vh",
+                        display: "flex",
+                        justifyContent: "center",
+                        alignItems: "center",
+                      }
                 }
-          }
-        >
-          {loadingPosts ? (
-            <Loader />
-          ) : (
-            <div className="app__posts">
-              {posts.map(({ id, post }) => (
-                <Post key={id} postId={id} user={user} post={post} />
-              ))}
+              >
+                {loadingPosts ? (
+                  <Loader />
+                ) : (
+                  <div className="app__posts">
+                    {posts.map(({ id, post }) => (
+                      <Post key={id} postId={id} user={user} post={post} />
+                    ))}
+                  </div>
+                )}
+              </div>
             </div>
+          ) : (
+            <></>
           )}
-        </div>
-      </div>
+        </Route>
+
+        <Route path="/dummygram/login">
+          <LoginScreen />
+        </Route>
+
+        <Route path="/dummygram/signup">
+          <SignupScreen />
+        </Route>
+
+        <Route path="*">
+          <h1 style={{ textAlign: "center", marginTop: "2rem" }}>
+            Page not found: 404
+          </h1>
+        </Route>
+      </Switch>
+
       <FaArrowCircleUp
         fill="#777"
         // stroke="30"
