+++ src/components/ImgUpload.jsx
@@ -187,6 +187,10 @@ function ImgUpload(props) {
         multiline
         rows={4}
         disabled={uploadingPost}
+        sx={{
+          backgroundColor: "white",
+          borderRadius: "8px",
+        }}
       />
       <AnimatedButton onClick={handleUpload} loading={uploadingPost}>
         Upload
