+++ src/main.jsx
@@ -4,6 +4,7 @@ import { createTheme, ThemeProvider } from "@mui/material/styles";
 import { SnackbarProvider, useSnackbar } from "notistack";
 import React from "react";
 import ReactDOM from "react-dom/client";
+import { BrowserRouter } from "react-router-dom";
 import App from "./App";
 import "./index.css";
 import LandingAnimation from "./components/LandingAnimation";
@@ -36,7 +37,9 @@ ReactDOM.createRoot(document.getElementById("root")).render(
           );
         }}
       >
-        <App />
+        <BrowserRouter>
+          <App />
+        </BrowserRouter>
       </SnackbarProvider>
     </ThemeProvider>
   </React.StrictMode>
