+++ src/reusableComponents/DialogBox.jsx
@@ -20,13 +20,30 @@ const DialogBox = (props) => {
       onClose={props.onClose}
       aria-labelledby="responsive-dialog-title"
     >
-      <DialogTitle id="responsive-dialog-title">{props.title}</DialogTitle>
+      <DialogTitle
+        id="responsive-dialog-title"
+        sx={{
+          fontWeight: "bold",
+          backgroundColor: "var(--bg-color)",
+          color: "var(--color)",
+        }}
+      >
+        {props.title}
+      </DialogTitle>
+
       <Divider />
-      <DialogContent>
+
+      <DialogContent sx={{ color: "var(--color)" }}>
         <DialogContentText>{props.children}</DialogContentText>
       </DialogContent>
-      <DialogActions>
-        <Button onClick={props.onClose}>Cancel</Button>
+
+      <DialogActions sx={{ backgroundColor: "var(--bg-color)" }}>
+        <Button
+          onClick={props.onClose}
+          sx={{ fontWeight: "bold", color: "var(--color)" }}
+        >
+          Cancel
+        </Button>
       </DialogActions>
     </Dialog>
   );
