+++ src/script.js
@@ -3,10 +3,10 @@ const options = {
   right: "32px", // default: '32px'
   left: "unset", // default: 'unset'
   time: "0.3s", // default: '0.3s'
-  mixColor: "#fff", // default: '#fff'
-  backgroundColor: "#fff", // default: '#fff'
-  buttonColorDark: "#100f2c", // default: '#100f2c'
-  buttonColorLight: "#fff", // default: '#fff'
+  // mixxColor: "#fff", // default: '#fff'
+  // backgroundColor: "#fff", // default: '#fff'
+  // buttonColorDark: "#100f2c", // default: '#100f2c'
+  // buttonColorLight: "#fff", // default: '#fff'
   saveInCookies: true, // default: true,
   label: "🌓", // default: ''
   autoMatchOsTheme: true, // default: true
@@ -22,9 +22,19 @@ document
   .querySelector("button.darkmode-toggle")
   .addEventListener("click", function () {
     if (document.body.classList.contains("darkmode--activated")) {
+      document.getElementsByClassName("app__header__img").item(0).style.filter =
+        "invert(100%)";
+      document.body.style.setProperty("--color", "white");
+      document.body.style.setProperty("--bg-color", "black");
+      window.document.body.style.setProperty("--val", 1);
       document.getElementById("scrollbar_style").innerHTML =
         "body::-webkit-scrollbar { width: 10px; } /* Track */ body::-webkit-scrollbar-track { background: black; } /* Handle */ body::-webkit-scrollbar-thumb { background: linear-gradient( 45deg, #f4f80c, #addb51, #a2b43a, #51c135, #5ce130, #1dfddf ); border-radius: 10px; }";
     } else {
+      document.getElementsByClassName("app__header__img").item(0).style.filter =
+        "invert(0%)";
+      document.body.style.setProperty("--color", "black");
+      document.body.style.setProperty("--bg-color", "white");
+      window.document.body.style.setProperty("--val", 0);
       document.getElementById("scrollbar_style").innerHTML =
         "body::-webkit-scrollbar { width: 10px; } /* Track */ body::-webkit-scrollbar-track { background: white; } /* Handle */ body::-webkit-scrollbar-thumb { background: linear-gradient( 45deg, #405de6,#5851db,#833ab4,#c13584,#e1306c,#fd1d1d ); border-radius: 10px; }";
     }
